Service ping complete
Updated Dockerfile - cleaned it up and changed the openssl request to
match the tutorial.

Added routes to the ssl keys to the config

Added a Server class which handles the connection to an http or https
server.

Updated index to use the Server class and to use the routes in the
tutorial.