Added ability for Routes to be generated dynamically
Routes are now defined by annotations on controller classes.
To define a route, create a class which will act as the subpath.
You can define a class with a @classRoute annotation to override the
default classname behavior.  Meaning that you can create any class
in the controllers directory, give it the classRoute annotation of "/"
and that class will be scanned on the next route cache build.

The Cache is only built once on deployment (to docker currently). It
must be deleted before it can be rebuilt currently. The cache is read
when a user connects and converted into handler methods and request
value checks.

@route - The route to use under the class route. (the classname or the
@classRoute annotation)

@json - expect to return json.
@xml - expect to return xml. (not implemented yet)
@html - expect to return html. (not implemented yet)
@text - expect to return text. (not implemented yet)
@method - GET, [GET, POST, ...]
@httpsOnly - Allows a GET request to be limited to https requests only.

@todo Considering the speed that this operates, we may be able to get
away with just rebuilding the cache every time the index is ran.