Add support for index-index,
resolve possible inf loop and add append func to Writer.

Cache Writing is temporary.
@todo - When index is spun up, build the cache like you do now. But,
instead of writing the routes to a file, have the build cache async
function return a promise. Index can add a "then" to the promise and
launch the servers when the routes are read from the controller files.
No cache files needed.  Since, the idea of a node app is a down and up,
stateless, will always be a fresh container for the user, kind of thing,
we would end up rebuilding the cache every time the app is spun anyway.